requirements.txt
python-telegram-bot==20.1
translate
requests
python-dotenv
apscheduler